Task 2/requirements.txt
psycopg[binary]>=3.2.2
python-dotenv==1.0.0
